ft(create): correct the test path
