+++ package.json
@@ -7,7 +7,7 @@
     "dev:start": "webpack server --mode development",
     "start": "webpack --mode production",
     "test": "react-scripts test",
-    "test:coverage": "react-scripts test --coverage a"
+    "test:coverage": "react-scripts test tests/ --coverage a"
   },
   "keywords": [],
   "author": "",
@@ -21,7 +21,8 @@
     "react-redux": "^7.2.8",
     "react-router-dom": "^6.3.0",
     "react-scripts": "^2.1.3",
-    "redux": "^4.1.2"
+    "redux": "^4.1.2",
+    "redux-form": "^8.3.8"
   },
   "devDependencies": {
     "@babel/core": "^7.17.9",
